% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_table.R
\docType{data}
\name{assoc.table}
\alias{assoc.table}
\title{Example of an Association Table}
\format{
A data frame with 443 rows and 4 columns.
}
\usage{
assoc.table
}
\description{
This association table is a data frame for use in the example for the rename_fasta function in this package RDPutils. It is identical to that produced in the example for the function assoc_repseq_IDs_with_otus_by_clstr.  The first column contains the machine names of representative sequences and the second column contains corresponding OTU names.
}
\keyword{RDPTools}
\keyword{datasets}
