% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_char_occurrences.R
\name{count_char_occurrences}
\alias{count_char_occurrences}
\title{Count Occurrences of a Character}
\usage{
count_char_occurrences(char, strng_x)
}
\arguments{
\item{char}{Character to search for}

\item{strng_x}{String that is searched}
}
\value{
An integer, the number of times a character occurs in a string.
}
\description{
Count the number of occurrences of a character in a string.
}
\examples{
my.strng <- "abcdefabcaa"
count_char_occurrences("a", strng_x=my.strng)
}
\author{
John Quensen
}
