% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_digits.R
\name{get_digits}
\alias{get_digits}
\title{Get digits}
\usage{
get_digits(otu.names)
}
\arguments{
\item{otu.names}{A vector of taxa names including underscores separating characters and digits}
}
\value{
A vector of digits
}
\description{
Get digits
}
\details{
Useful in reformating OTU/taxa names, sometimes necessary to merge phyloseq objects.
}
