% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repseq_IDs_from_fasta.R
\name{get_repseq_IDs_from_fasta}
\alias{get_repseq_IDs_from_fasta}
\title{Get Representative Sequence Names}
\usage{
get_repseq_IDs_from_fasta(repseq_file = "all_seq_complete.clust_rep_seqs.fasta")
}
\arguments{
\item{repseq_file}{The file name of the combined fasta file containing the representative sequences.}
}
\value{
A vector of the machine names of the representative sequences.
}
\description{
This function retrieves a vector of the representative sequence machine names from a fasta file of representative sequences.
}
\details{
The function reads the fasta file of representative sequences from disk.
}
\examples{
repseq.file <- system.file("extdata", "all_seq_complete.clust_rep_seqs.fasta", package="RDPutils")
rep.seqs <- get_repseq_IDs_from_fasta(repseq_file = repseq.file)
rep.seqs
}
\references{
The web-based tool for retrieving representative sequences is here:
http://pyro.cme.msu.edu/
 
A command line version of the tool is included in RDPTools, available on GitHub:
https://github.com/rdpstaff
}
\author{
John Quensen
}
\keyword{RDPTols}
