% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_itagger_otutab_taxa.R
\name{import_itagger_otutab_taxa}
\alias{import_itagger_otutab_taxa}
\title{Import iTagger File}
\usage{
import_itagger_otutab_taxa(in_file)
}
\arguments{
\item{in_file}{otu.tax.tsv file from iTagger}
}
\value{
A phyloseq object with OTU and taxonomy tables.
}
\description{
Converts the tab-delimited iTagger otu.tax.tsv file into a phyloseq object with otu_table and tax_table.
}
\details{
The iTagger otu.tax.tsv file is similar to legacy QIIME format. This function expects six ranks: Kingdom, Phylum, Class, Order, Family, and Genus. The taxonomy field in otu.tax.tsv is parsed on semicolons.
}
\examples{
##---- Not run. ----
##-- expt <- import_itagger_otutab_taxa(in_file = "iTagger_otutab_taxa_file.txt")
}
\author{
John Quensen
}
