\name{import_otutab_taxa}
\alias{import_otutab_taxa}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Import USEARCH otutab_taxa File into phyloseq
}
\description{
Converts USEARCH's usearch_global output option otutabout to a phyloseq object with otu_table and tax_table.
}
\usage{
import_otutab_taxa(in_file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{in_file}{
An otutab_taxa file generated by USEARCH 8.1+
}
}
\details{
The USEARCH usearch_global function has an option to output an OTU table with taxonomy assignments in a tab-delimited text file. The output resembles the legacy QIIME format. This function imports such a file into a phyloseq object with otu_table and tax_table.
}
\value{
A phyloseq object with otu_table and tax_table.
}

\author{
John Quensen
}

\examples{
##---- Not run. ----
##-- expt <- import_otutab_taxa(in_file = "otutab_taxa_03.txt")

}

\keyword{ USEARCH }

