% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2_taxonomy.R
\name{import_qiime2_taxonomy}
\alias{import_qiime2_taxonomy}
\title{Import QIIME2 Taxonomy File}
\usage{
import_qiime2_taxonomy(in_file)
}
\arguments{
\item{in_file}{}
}
\value{
A phyloseq tax_table
}
\description{
Converts a tab-delimited taxonomy table made with QIIME2 to a phyloseq tax_table object.
}
\references{
Bolyen E, Rideout JR, et al., 2019. Reproducible, interactive, scalable and extensible microbiome data science using QIIME 2. Nature Biotechnology 37: 852–857. https://doi.org/10.1038/s41587-019-0209-9
}
\author{
John Quensen
}
