% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_usearch_biom_file.R
\name{import_usearch_biom_file}
\alias{import_usearch_biom_file}
\title{Import a USEARCH biom File}
\usage{
import_usearch_biom_file(in_file, ...)
}
\arguments{
\item{in_file}{A biom file produced with USEARCH's usearch_global command.}

\item{...}{Other objects accepted by phyloseq, e.g. sample data table, representative sequences.}
}
\value{
An experiment level phyloseq object (i.e. with at least an otu_table).
}
\description{
Converts a biom file produced with USEARCH's usearch_global command into a phyloseq object.
}
\details{
This function will import the OTU table contained in the USEARCH biom file, and the taxonomy table if it is present. The function will also accept arguments for sample data table, reference sequences, and a phyloseq class taxonomy table from another source, if one is not present in the biom file.
}
\examples{
##---- Not run. ----
##-- expt <- import_usearch_biome(infile = "otutab_taxa_03.json")
}
\author{
John Quesen
}
