\name{otu}
\docType{data}
\alias{otu}
\title{Example of an OTU Table}
\description{
  This OTU table is for use with the examples in this package RDPutils. There are 4 samples in rows by 443 OTUs in columns.  It was made by clustering 4 samples of 250 partial 16S rRNA gene sequences at a distance of 0.03.
}
\usage{otu}
\format{A data frame with 4 rows and 443 columns.}

\keyword{datasets}
