% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_taxa_names.R
\name{reformat_taxa_names}
\alias{reformat_taxa_names}
\title{Reformat Taxa Names}
\usage{
reformat_taxa_names(p, otu_format = "R")
}
\arguments{
\item{p}{A phyloseq object including taxa}

\item{otu_format}{A string specifying format}
}
\value{
A phyloseq object with reformatted taxa names
}
\description{
Reformat Taxa Names
}
\details{
Format string is either "R" (the default) or "biom", corresponding to the formats returned by RDPTools.
}
