% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_fasta.R
\name{rename_fasta}
\alias{rename_fasta}
\title{Rename Representative Sequences}
\usage{
rename_fasta(in_file = "names_trimmed.fasta", out_file = "renamed.fasta", rename_table)
}
\arguments{
\item{in_file}{The name of the fasta file containing representative sequences with their machine names.}

\item{out_file}{The name of the fasta file containing representative sequences with their OTU names.}

\item{rename_table}{A data frame associating the machine names of representative sequences with their corresponding OTU names.}
}
\description{
This function renames representative sequences from their machine names to their OTU names.
}
\details{
The output of the function assoc_repseq_IDs_with_otus_by_fasta or assoc_repseq_IDs_with_otus_by_clstr  can be used as the rename_table.
Before representative sequences can be renamed with this function, their IDs have to be shortened to include only their machine names with the function trim_fasta_names.
}
\examples{
in.file <- system.file("extdata", "names_trimmed.fasta", package="RDPutils")
data(assoc.table)
rename_fasta(in_file = in.file, out_file = "renamed.fasta", rename_table = assoc.table)
}
\seealso{
[assoc_repseq_IDs_with_otus_by_clstr()]
[assoc_repseq_IDs_with_otus_by_fasta()]
[trim_fasta_names()]
}
\author{
John Quensen
}
\keyword{RDPTools}
