\name{assoc.table}
\docType{data}
\alias{assoc.table}
\title{Example of an Association Table}
\description{
This association table is a data frame for use in the example for the rename_fasta function in this package RDPutils. It is identical to that produced in the example for the function assoc_repseq_IDs_with_otus_by_clstr.  The first column contains the machine names of representative sequences and the second column contains corresponding OTU names.
}
\usage{assoc.table}
\format{A data frame with 443 rows and 4 columns.}

\keyword{datasets}
