% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_otutab_taxa.R
\name{import_otutab_taxa}
\alias{import_otutab_taxa}
\title{Import USEARCH otutab_taxa File into phyloseq}
\usage{
import_otutab_taxa(in_file)
}
\arguments{
\item{in_file}{An otutab_taxa file generated by USEARCH 8.1+}
}
\value{
A phyloseq object with otu_table and tax_table.
}
\description{
Converts USEARCH's usearch_global output option otutabout to a phyloseq object with otu_table and tax_table.
}
\details{
The USEARCH usearch_global function has an option to output an OTU table with taxonomy assignments in a tab-delimited text file. The output resembles the legacy QIIME format. This function imports such a file into a phyloseq object with otu_table and tax_table.
}
\examples{
##---- Not run. ----
##-- expt <- import_otutab_taxa(in_file = "otutab_taxa_03.txt")
}
\author{
John Quensen
}
\keyword{USEARCH}
