\name{import_otutab_taxa}
\alias{import_otutab_taxa}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Import USEARCH otutab_taxa File into phyloseq
}
\description{
Converts USEARCH's usearch_global output option otutabout to a phyloseq object with otu_table and tax_table.
}
\usage{
import_otutab_taxa(in_file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{in_file}{
An otutab_taxa file generated by USEARCH 8.1+
}
}
\details{
The USEARCH usearch_global function has an option to output an OTU table with taxonomy assignments in a tab-delimited text file. The output resembles the legacy QIIME format. This function imports such a file into a phyloseq object with otu_table and tax_table.
}
\value{
A phyloseq object with otu_table and tax_table.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
John Quensen
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Not run. ----
##-- expt <- import_otutab_taxa(in_file = "otutab_taxa_03.txt")

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ USEARCH }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
