% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_utax_file.R, R/make_framebot_tax_table.R
\name{import_utax_file}
\alias{import_utax_file}
\alias{make_framebot_tax_table}
\title{Import utax Taxonomy File}
\usage{
import_utax_file(in_file, confidence)

import_utax_file(in_file, confidence)
}
\arguments{
\item{in_file}{The utaxout result from USEARCH's utax command.}

\item{confidence}{The confidence level to use in assigning taxonomic categories.}
}
\value{
A phyloseq tax_table.

A phyloseq tax_table.
}
\description{
Converts the tab-delimited output of USEARCH's utax command to a phyloseq tax_table object. The confidence level for taxonomic assignment is chosen on import.

Converts the tab-delimited output of USEARCH's utax command to a phyloseq tax_table object. The confidence level for taxonomic assignment is chosen on import.
}
\details{
The default confidence level is 0.8.

The default confidence level is 0.8.
}
\examples{
##---- Not run. ----
##-- tax.table <- import_utax_file(in_file="utax_result.txt", confidence = 0.8)
##---- Not run. ----
##-- tax.table <- import_utax_file(in_file="utax_result.txt", confidence = 0.8)
}
\author{
John Quensen
}
\keyword{USEARCH}
