% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_otu_names.R
\name{make_otu_names}
\alias{make_otu_names}
\title{Make OTU Names}
\usage{
make_otu_names(otu_nums, otu_format="R")
}
\arguments{
\item{otu_nums}{A vector of integers.}

\item{otu_format}{When equal to "R" (default) OTU names have the form "OTUxxxnn."  When equal to "biom", OTU names have the form "cluster_nn."}
}
\value{
A vector of character strings of the form "OTU_xxxnn"" or "cluster_nn."
}
\description{
This function takes a vector of integers and makes a vector of OTU names.
}
\details{
This function is used by the function clstr2otu to name the columns of the OTU table it returns, but make_otu_names can be used separately if desired.
}
\examples{
n <- c(1:10)
otu.names <- make_otu_names(n)
otu.names
}
\author{
John Quensen
}
\keyword{RDPTools}
