% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_model_seqs.R
\name{remove_model_seqs}
\alias{remove_model_seqs}
\title{Remove Model Sequences}
\usage{
remove_model_seqs(in_file, out_file = in_file)
}
\arguments{
\item{in_file}{The name of a combined fasta file with model sequences.}

\item{out_file}{The name of the corresponding fasta file written to disk without model sequences.}
}
\value{
The function returns a message that it has completed.
}
\description{
This function removes model sequences from a combined fasta file.
}
\details{
This function operates on files.  It is not normally assigned to a variable.  By default, the input file is overwritten.  If no model sequence is found in the input file, a message to that effect is returned.
RDP aligns 16S rRNA gene sequences using the Infernal aligner with model sequences for Bacteria or Archaea.  A model sequence is introduced into each alignment.  These model sequences are required by the cluster tool and the merge alignment tool.  If a model sequence is present in the fasta file of representative sequences, it needs to be removed before the sequences can be treed with FastTree; this function provides the means to do so.
}
\examples{
in.file <- system.file("extdata", "all_seq_complete.clust_rep_seqs.fasta", package="RDPutils")
remove_model_seqs(in_file=in.file, out_file = file2tree.fasta)
}
\author{
John Quensen
}
\keyword{RDPTools}
